네이밍 보완, 함수통합, APP컴포넌트 분리
