+++ src/Buttons.jsx
@@ -1,13 +1,13 @@
 import React from 'react';
 import Button from './Button';
 
-function Buttons({ onButtonClick }) {
+function Buttons({ onNumberClick }) {
   return (
     <p>
       {[1, 2, 3, 4, 5].map((i) => (
         <Button
           key={i}
-          onButtonClick={() => onButtonClick(i)}
+          onNumberClick={() => onNumberClick(i)}
         >
           {i}
         </Button>
