+++ src/index.jsx
@@ -1,29 +1,8 @@
-import React, { useState } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
-import Page from './Page';
 
-function App() {
-  const [state, setState] = useState({
-    count: 0,
-  });
+import App from './App';
 
-  const { count } = state;
-
-  function handleClick(i) {
-    setState({
-      count: count + i,
-    });
-  }
-
-
-  return (
-    <Page
-      count={count}
-      onClick={handleClick}
-      onButtonClick={handleClick}
-    />
-  );
-}
 
 ReactDOM.render(
   <App />,
